+++ src/Snake.cpp
@@ -1,9 +1,36 @@
 // Snake.cpp
 #include <iostream>
+#include <queue>
 #include <SFML\Graphics.hpp>
-#include "GeneralFunctions.h"
 #include "Snake.h"
 #include "SnakeFood.h"
+
+// Get a new coordinate position based on snake direction
+sf::Vector2f CalculateNewPosition(int direction, sf::Vector2f position)
+{
+    if (direction == 0)
+        return position;
+    if (direction == 1)
+        position.x -= 25;
+    if (direction == 2)
+        position.y -= 25;
+    if (direction == 3)
+        position.y += 25;
+    if (direction == 4)
+        position.x += 25;
+    return position;
+}
+
+// Test for collision between two object positions
+bool GlobalCollision(sf::Vector2f object1Position, sf::Vector2f object2Position)
+{
+    if (object1Position.x != object2Position.x)
+        return 0;
+    if (object1Position.y != object2Position.y)
+        return 0;
+
+    return 1;
+}
 
 // Check keyboard for new direction of snake
 void Snake::CheckDirection()
