+++ src/Snake.h
@@ -1,9 +1,9 @@
 // Snake.h
 #ifndef SNAKE_H
 #define SNAKE_H
-#include <queue>
-#include "SnakeFood.h"
 
+sf::Vector2f CalculateNewPosition(int direction, sf::Vector2f position);
+bool GlobalCollision(sf::Vector2f object1Position, sf::Vector2f object2Position);
 
 class Snake
 {
